PHP 8.1 fixes

+++ galette/lib/Galette/IO/PdfMembersCards.php
@@ -201,7 +201,7 @@ class PdfMembersCards extends Pdf
 
             // Compute card position on page
             $col = $nb_card % $this->nbcol;
-            $row = ($nb_card / $this->nbcol) % $this->nbrow;
+            $row = (int)(($nb_card / $this->nbcol)) % $this->nbrow;
             // Set origin
             $x0 = $this->xorigin + $col * (round($this->wi) + round($this->hspacing));
             $y0 = $this->yorigin + $row * (round($this->he) + round($this->vspacing));
